requirements.txt
flask==2.3.3
cohere==4.37
python-dotenv==1.0.0
gunicorn==21.2.0 